requirements.txt
Flask
Pillow
opencv-python
git+ssh://git@github.com/wobeng/Zappa.git@master